+++ src/genome/tf/motif/__init__.py
@@ -22,6 +22,7 @@ from genome.tf.motif.jaspar import (
     TransfacError,
     parse_transfac,
 )
+from genome.tf.motif.mixin import REGION_HIT_PROVENANCE, MotifScanMixin
 from genome.tf.motif.motif import (
     BASES,
     DEFAULT_SEQUENCE_NAME,
@@ -61,6 +62,7 @@ __all__ = [
     "JASPAR_RELEASES",
     "JASPAR_TAX_GROUPS",
     "MIN_MOTIF_LENGTH",
+    "REGION_HIT_PROVENANCE",
     "SLURM_CPU_VARS",
     "UNIFORM_BACKGROUND",
     "AmbiguousBaseIdError",
@@ -72,6 +74,7 @@ __all__ = [
     "Motif",
     "MotifComparison",
     "MotifNotFoundError",
+    "MotifScanMixin",
     "MotifSet",
     "RaggedComparisonError",
     "TransfacError",